# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Roblox Vector Search API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "summary": "Get all jobs and their status with statistics",
        "operationId": "getAllJobs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of jobs to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of jobs to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter jobs by status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "running",
                "completed",
                "failed"
              ]
            }
          },
          {
            "name": "command",
            "in": "query",
            "description": "Filter jobs by command",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs with statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobs": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "command": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "pending",
                                  "running",
                                  "completed",
                                  "failed"
                                ]
                              },
                              "progress": {
                                "type": "object",
                                "properties": {
                                  "current": {
                                    "type": "number"
                                  },
                                  "total": {
                                    "type": "number"
                                  },
                                  "message": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "current",
                                  "total"
                                ],
                                "additionalProperties": false
                              },
                              "result": {},
                              "error": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "started_at": {
                                "type": "string"
                              },
                              "completed_at": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "command",
                              "status",
                              "created_at"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "stats": {
                          "type": "object",
                          "properties": {
                            "pending": {
                              "type": "number"
                            },
                            "running": {
                              "type": "number"
                            },
                            "completed": {
                              "type": "number"
                            },
                            "failed": {
                              "type": "number"
                            },
                            "total": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "pending",
                            "running",
                            "completed",
                            "failed",
                            "total"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "jobs",
                        "stats"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "success": true,
                  "data": {
                    "jobs": [
                      {
                        "id": "string",
                        "command": "string",
                        "status": "string",
                        "progress": {
                          "current": 0,
                          "total": 0,
                          "message": "string"
                        },
                        "result": null,
                        "error": null,
                        "created_at": "string",
                        "started_at": "string",
                        "completed_at": "string"
                      }
                    ],
                    "stats": {
                      "pending": 0,
                      "running": 0,
                      "completed": 0,
                      "failed": 0,
                      "total": 0
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Global error response",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Error message"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "data": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}